refactor(vardump): APIを整理
