+++ packages/engine/src/wwa_vardump/index.ts
@@ -1,17 +1,11 @@
 import { $qs } from "../wwa_util";
+import * as UserVariableListSection from "./user-variable/user-variable-list-section";
 
-import * as UserVariableCard from "./user-variable-card";
-import * as UserVariableLabel from "./user-variable-label";
-import * as UserVariableList from "./user-variable-list";
-import * as UserVariableListSection from "./user-variable-list-section";
-
-export * from "./apis";
-export {
-  UserVariableCard,
-  UserVariableLabel,
-  UserVariableList,
-  UserVariableListSection,
-};
+export * as UserVariableCard from "./user-variable/user-variable-card";
+export * as UserVariableLabel from "./user-variable/user-variable-label";
+export * as UserVariableList from "./user-variable/user-variable-list";
+export * as NumberedUserVariable from "./numbered-user-variable";
+export { UserVariableListSection };
 
 export const CLASS_NAME = "wwa-vardump-wrapper";
 
@@ -23,11 +17,7 @@ export function setup(dumpElmQuery: string): HTMLElement | null {
   }
 
   element.classList.add(CLASS_NAME);
-  /*
-  element.appendChild(
-    UserVariableListSection.createElement({ kind: "named" })
-  );
-  */
+  element.appendChild(UserVariableListSection.createElement({ kind: "named" }));
   element.appendChild(
     UserVariableListSection.createElement({ kind: "numbered" })
   );
